# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/basics/inspectors": {
            "get": {
                "description": "This endpoint retrieves ticket inspector reports from the database within a specified time range.\nIt supports filtering by start and end timestamps,\nand checks if the data has been modified since the last request using the \"If-Modified-Since\" header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basics"
                ],
                "summary": "Retrieve information about ticket inspector reports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End timestamp (RFC3339 format)",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Station ID to filter inspectors for a specific station",
                        "name": "station",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/utils.TicketInspectorResponse"
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified: The data has not changed since the If-Modified-Since header."
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity: The station does not belong to the network.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Accepts a JSON payload with details about a ticket inspector's current location.\nThis endpoint validates the provided data, processes necessary computations for linking stations and lines,\ninserts the data into the database, and triggers an update to the risk model used in operational analysis.\nIf the 'timestamp' field of the body is not provided, the current UTC time truncated to the nearest minute is used automatically.\nThe endpoint is rate limited per IP address, and reports classified as spam are rejected with 403.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basics"
                ],
                "summary": "Submit ticket inspector data",
                "parameters": [
                    {
                        "description": "Data about the inspector's location and activity, including an optional 'timestamp' in RFC3339 format",
                        "name": "inspectorData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.InspectorRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully processed and inserted the inspector data with computed linkages and risk model updates.",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Missing or incorrect parameters provided."
                    },
                    "403": {
                        "description": "Forbidden: The report was classified as spam.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity: The report names a station, direction or line the network does not contain.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests: The request has been rate limited."
                    },
                    "500": {
                        "description": "Internal Server Error: Error during data processing or database insertion."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "utils.InspectorRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "is always null or 98111116 (ASCII for \"BOT\") when getting data from the telegram bot",
                    "type": "integer"
                },
                "directionId": {
                    "type": "string"
                },
                "line": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "stationId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "utils.ResponseData": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "direction": {
                    "$ref": "#/definitions/utils.Station"
                },
                "line": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "station": {
                    "$ref": "#/definitions/utils.Station"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "utils.Station": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/utils.Coordinates"
                },
                "id": {
                    "type": "string"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "utils.TicketInspectorResponse": {
            "type": "object",
            "properties": {
                "direction": {
                    "$ref": "#/definitions/utils.Station"
                },
                "isHistoric": {
                    "type": "boolean"
                },
                "line": {
                    "description": "String is used so that it can easily be handled by the frontend",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "station": {
                    "$ref": "#/definitions/utils.Station"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}